resolviendo bugs
