+++ config/filesystems.php
@@ -2,7 +2,7 @@
 
 return [
 
-    /*
+   /*
     |--------------------------------------------------------------------------
     | Default Filesystem Disk
     |--------------------------------------------------------------------------
@@ -13,9 +13,9 @@ return [
     |
     */
 
-    'default' => env('FILESYSTEM_DRIVER', 'local'),
+   'default' => env('FILESYSTEM_DRIVER', 'local'),
 
-    /*
+   /*
     |--------------------------------------------------------------------------
     | Default Cloud Filesystem Disk
     |--------------------------------------------------------------------------
@@ -26,9 +26,9 @@ return [
     |
     */
 
-    'cloud' => env('FILESYSTEM_CLOUD', 's3'),
+   'cloud' => env('FILESYSTEM_CLOUD', 's3'),
 
-    /*
+   /*
     |--------------------------------------------------------------------------
     | Filesystem Disks
     |--------------------------------------------------------------------------
@@ -41,28 +41,28 @@ return [
     |
     */
 
-    'disks' => [
+   'disks' => [
 
-        'local' => [
-            'driver' => 'local',
-            'root' => storage_path('app'),
-        ],
+      'local' => [
+         'driver' => 'local',
+         'root' => storage_path('app'),
+      ],
 
-        'public' => [
-            'driver' => 'local',
-            'root' => storage_path('app/public'),
-            'url' => env('APP_URL').'/storage',
-            'visibility' => 'public',
-        ],
+      'public' => [
+         'driver' => 'local',
+         'root' => storage_path('app/public'),
+         'url' => env('APP_URL') . '/storage/app/public',
+         'visibility' => 'public',
+      ],
 
-        's3' => [
-            'driver' => 's3',
-            'key' => env('AWS_KEY'),
-            'secret' => env('AWS_SECRET'),
-            'region' => env('AWS_REGION'),
-            'bucket' => env('AWS_BUCKET'),
-        ],
+      's3' => [
+         'driver' => 's3',
+         'key' => env('AWS_KEY'),
+         'secret' => env('AWS_SECRET'),
+         'region' => env('AWS_REGION'),
+         'bucket' => env('AWS_BUCKET'),
+      ],
 
-    ],
+   ],
 
 ];
