+++ resources/assets/js/components/Estudiantes/ProfileComponent.vue
@@ -11,9 +11,9 @@
                      <v-layout xs12>
                         <v-flex xs5 class="text-xs-center">
                            <img
-                              :src="estudiante.foto!='none'?estudiante.foto:'/storage/imagen/default-user.png'"
+                              :src="estudiante.foto!='none'?estudiante.foto:'/storage/app/public/imagen/default-user.png'"
                               height="125"
-                           >
+                           />
                            <v-flex
                               v-if="edit"
                               class="text-xs-center text-sm-center text-md-center text-lg-center"
@@ -31,7 +31,7 @@
                                  ref="image"
                                  accept="image/*"
                                  @change="onFilePicked"
-                              >
+                              />
                            </v-flex>
                         </v-flex>
                         <v-flex xs7>
@@ -213,8 +213,7 @@ export default {
                };
                axios.post("/store", data, config).then(function(res) {
                   if (res.data.name != "error") {
-                     self.estudiante.foto =
-                        "/storage" + res.data.name.substr(6);
+                     self.estudiante.foto = "/storage/app/" + res.data.name;
                   } else {
                      self.$toast.error("No se pudo cargar esta imagen", {
                         y: "top",
